ArloPilot v1.5: Virtual camera tiles update.
What's fixed?
Last image polling not being properly scheduled during device creation.

What's new?
ArloPilot now checks to see if the image returned from the cloud is different than the last check.  This is accomplished during polling for last image by computing an MD5 hash on the image.  If the poll returns an image with the same hash it is not stored by SmartThings.